+++ src/features/MotionList/ActionList/ListItem.tsx
@@ -4,7 +4,9 @@ import classNames from 'classnames';
 import React, { memo } from 'react';
 
 import ListItem from '@/components/ListItem';
+import { useGlobalStore } from '@/store/global';
 import { MotionAnimation } from '@/types/touch';
+import { fetchWithProgress } from '@/utils/fetch';
 
 interface ActionListItemProps {
   item: MotionAnimation;
@@ -12,7 +14,7 @@ interface ActionListItemProps {
 
 const useStyles = createStyles(({ css, token }) => ({
   listItem: css`
-    height: 56px;
+    height: 64px;
     margin-block: 2px;
 
     font-size: ${token.fontSize}px;
@@ -24,14 +26,23 @@ const useStyles = createStyles(({ css, token }) => ({
 
 const TouchActionListItem = memo<ActionListItemProps>(({ item }) => {
   const { styles } = useStyles();
+  const viewer = useGlobalStore((s) => s.viewer);
 
   return (
     <ListItem
       key={item.id}
       className={classNames(styles.listItem)}
-      description={item.description.slice(0, 120)}
+      description={item.description.slice(0, 40)}
       avatar={<Avatar src={item.avatar} shape="square" />}
-      showAction={true}
+      showAction={false}
+      onClick={async () => {
+        if (item.url) {
+          const blob = await fetchWithProgress(item.url);
+          const url = window.URL.createObjectURL(blob);
+
+          viewer.model?.loadFBX(url);
+        }
+      }}
       title={item.name}
       active={false}
     />
