--- src/features/MotionList/style.ts
@@ -1,9 +1,0 @@
-import { createStyles } from 'antd-style';
-
-export const useStyles = createStyles(({ css }) => ({
-  list: css`
-    position: relative;
-    overflow: scroll;
-    height: 100%;
-  `,
-}));
